requirements.txt

pypdf
sentence-transformers
faiss-cpu
transformers
accelerate
torch
langchain
